+++ tests/test_pact.py
@@ -1,10 +1,10 @@
 import pytest
-from contract.pact import Pact
+from contract.pact import Contract
 from .conftest import client_http
 
 
 def test_pact_green_path():
-    pact = Pact('tests/pacts/pact1.json')
+    contract = Contract('tests/contracts/pact1.json')
 
     expected = {
         "user": {"name": "felipe", "lastname": "volpone"},
@@ -14,13 +14,13 @@ def test_pact_green_path():
     # fake client. it will fake a http call
     client_api = client_http(200, expected)
 
-    pact.select('get one user').call(client_api).assert_it()
+    contract.select('get one user').call(client_api).assert_it()
 
 
 def test_pact_select_interaction_doesnot_exists():
     # fake client. it will fake a http call
     client_api = client_http(200, {})
 
-    pact = Pact('tests/pacts/pact1.json')
+    contract = Contract('tests/contracts/pact1.json')
     with pytest.raises(Exception):
-        pact.select('i dont exist').call(client_api).assert_it()
+        contract.select('i dont exist').call(client_api).assert_it()