+++ CHANGES.rst
@@ -4,4 +4,4 @@ Changelog
 0.1.0
 ~~~~~~
 
-* First release!
+* First release! API to create the contract and to write a test with it